/etc/scion config volume option
add an option that controls the handling of SCION related configuration files (contents of /etc/scion dir ).
Now users can choose whether to bake them into the docker images(as it used to be) or have them bind-mounted from a shared folder on the host. This can be used to reduce image build times.